requirements.txt
langchain
langchain-community
langchain-core
langchain-ollama
chromadb
pypdf
streamlit
python-dotenv
sentence-transformers
ollama
langchain-text-splitters
rank_bm25
